How to read a file
To read a file, you must open it with "r" second parameter
fopen($file, "r")
and then use the function fread()
this function takes two arguments, the file name, and the number of
characters to read. If you want to read the whole file, you can use
the filesize() function in the second argument.
It always good to hold the return of the fread() in a variable